correct be implementation

+++ src/main/java/prova/demo/repository/ClienteRepository.java
@@ -3,5 +3,5 @@ package prova.demo.repository;
 import org.springframework.data.jpa.repository.support.JpaRepositoryImplementation;
 import prova.demo.entity.Cliente;
 
-public interface ClienteRepository extends JpaRepositoryImplementation<Cliente, Long> {
+public interface ClienteRepository extends JpaRepositoryImplementation<Cliente, Integer> {
 }
