+++ src/main/java/prova/demo/repository/TecnicoRepository.java
@@ -3,5 +3,5 @@ package prova.demo.repository;
 import org.springframework.data.jpa.repository.support.JpaRepositoryImplementation;
 import prova.demo.entity.Tecnico;
 
-public interface TecnicoRepository extends JpaRepositoryImplementation<Tecnico, Long> {
+public interface TecnicoRepository extends JpaRepositoryImplementation<Tecnico, Integer> {
 }
